github-guardian-backend/requirements.txt
fastapi
uvicorn[standard]
PyGithub
python-dotenv
celery
redis
sqlalchemy
alembic
pydantic-settings
pyyaml
httpx
docker
slack-sdk
gunicorn
